backend/requirements.txt
colorama==0.4.6
coverage==7.6.12
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pytesseract==0.3.13
pytest==8.3.5
pytest-cov==6.0.0
tomli==2.2.1